refactor(admin): create campaign-participations tab
