+++ admin/app/components/users/user-detail-personal-information.hbs
@@ -14,13 +14,6 @@
   />
 </section>
 
-<section class="page-section">
-  <Users::UserDetailPersonalInformation::CampaignParticipations
-    @participations={{@user.participations}}
-    @removeParticipation={{@removeParticipation}}
-  />
-</section>
-
 <ConfirmPopup
   @message="Êtes-vous sûr de vouloir dissocier ce prescrit ?"
   @title="Confirmer la dissociation"
