+++ admin/app/controllers/authenticated/users/get/information.js
@@ -53,18 +53,6 @@ export default class UserInformationController extends Controller {
     }
   }
 
-  @action
-  async removeParticipation(campaignParticipation) {
-    try {
-      await campaignParticipation.deleteRecord();
-      await campaignParticipation.save();
-      await this.model.participations.reload();
-      this.notifications.success('La participation du prescrit a été supprimée avec succès.');
-    } catch (e) {
-      this.notifications.error('Une erreur est survenue lors de la suppression de la participation.');
-    }
-  }
-
   _handleResponseError(errorResponse, identityProvider) {
     const { errors } = errorResponse;
