+++ admin/app/router.js
@@ -46,6 +46,7 @@ Router.map(function () {
       this.route('get', { path: '/:user_id' }, function () {
         this.route('information', { path: '/' });
         this.route('profile');
+        this.route('campaign-participations', { path: '/participations' });
       });
     });
