+++ admin/app/routes/authenticated/users/get/campaign-participations.js
@@ -1,0 +1,9 @@
+import Route from '@ember/routing/route';
+
+export default class UserProfileRoute extends Route {
+  async model() {
+    const user = this.modelFor('authenticated.users.get');
+    await user.hasMany('participations').reload();
+    return user;
+  }
+}
