+++ admin/app/templates/authenticated/users/get/campaign-participations.hbs
@@ -1,0 +1,6 @@
+<section class="page-section">
+  <Users::CampaignParticipations
+    @participations={{@model.participations}}
+    @removeParticipation={{this.removeParticipation}}
+  />
+</section>
