+++ admin/app/templates/authenticated/users/get/information.hbs
@@ -3,5 +3,4 @@
   @removeAuthenticationMethod={{this.removeAuthenticationMethod}}
   @addPixAuthenticationMethod={{this.addPixAuthenticationMethod}}
   @reassignAuthenticationMethod={{this.reassignAuthenticationMethod}}
-  @removeParticipation={{this.removeParticipation}}
 />
